Mirror page tabs with entry tabs
