+++ src/components/entry/EntryTabs.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components/macro'
 
-import { PageTabs, Tab, TabList, TabPanel, TabPanels } from '../ui/PageTabs'
+import { EntryTab, Tab, TabList, TabPanel, TabPanels } from '../ui/EntryTabs'
 import EntryDetails from './EntryDetails'
 import EntryReviews from './EntryReviews'
 
@@ -31,25 +31,7 @@ export const TextContent = styled.article`
   }
 `
 
-// TODO: create /ui/EntryTabs.js, and move all UI stuff into it. Mirror the PageTabs
 // Rename Overview and Reviews
-const EntryTab = styled(PageTabs)`
-  [data-reach-tab-list] {
-    height: 40px;
-    [data-reach-tab] {
-      display: flex;
-
-      font-size: 14px;
-      border-top: none;
-
-      border-bottom: 3px solid ${({ theme }) => theme.secondaryBackground};
-      &[data-selected] {
-        color: ${({ theme }) => theme.orange};
-        border-bottom-color: ${({ theme }) => theme.orange};
-      }
-    }
-  }
-`
 
 const EntryTabs = () => {
   console.log('tabs')
